Nowy widok admina
Dodano widok admina
Dodano akcję wszystkie rezerwacje w systemie

+++ app/controllers/AllReservationsCtrl.class.php
@@ -9,7 +9,7 @@ use core\Utils;
 
 
 
-class MyReservationsCtrl {
+class AllReservationsCtrl {
     
     private $id_user;
     private $reservations;
@@ -29,20 +29,19 @@ class MyReservationsCtrl {
     public function getParams(){
         $this->id_user = \core\SessionUtils::load("id_user", $keep = true);
         $this->role = \core\SessionUtils::load("role", $keep = true);
-        $this->id_reservation = \core\ParamUtils::getFromCleanURL(1, false);
+        $this->id_reservation = \core\ParamUtils::getFromRequest('reservation');
         
     }
     
-    public function action_my_reservations() {   
+    public function action_all_reservations() {   
         $this->getParams();
         $this->getMyReservations();
         $this->generateView();
     }
     
-    public function action_delete_reservation() {   
-
+    public function action_delete_admin_reservation() {   
+       if($this->validate()){
            print($this->id_reservation);
-           
             try {
             $this->reservations = App::getDB()->delete("reservation", [
                 "AND" => [
@@ -54,9 +53,11 @@ class MyReservationsCtrl {
                 if (App::getConf()->debug)
                     Utils::addErrorMessage($e->getMessage());
             }
-        
-        $this->action_my_reservations();
+        } 
+        $this->action_all_reservations();
     }
+    
+    
    
     public function getMyReservations() {  
 
@@ -64,7 +65,8 @@ class MyReservationsCtrl {
             try {
             $this->reservations = App::getDB()->select("reservation", 
                 [
-                    "[>]service" => "id_service"
+                    "[>]service" => "id_service",
+                    "[>]user" => "id_user"
                 ],[
                     
                     "reservation.id_reservation", 
@@ -72,10 +74,10 @@ class MyReservationsCtrl {
                     "reservation.id_service",
                     "reservation.id_user",
                     "service.service_name",
-                    "service.service_time"
-                ], [
-                    "reservation.id_user" => $this->id_user
-        ]);
+                    "service.service_time",
+                    "user.name",
+                    "user.phone_number",
+                ]);
             } catch (\PDOException $e) {
                 Utils::addErrorMessage('Wystąpił błąd podczas pobierania rekordów');
                 if (App::getConf()->debug)
@@ -92,6 +94,6 @@ class MyReservationsCtrl {
         App::getSmarty()->assign("reservations", $this->reservations);
         App::getSmarty()->assign("action_url",App::getConf()->action_url);      
         App::getSmarty()->assign("app_url",App::getConf()->app_root);        
-        App::getSmarty()->display("my_reservations.tpl");    
+        App::getSmarty()->display("all_reservations.tpl");    
     }
 }
